feat: big app refactor - changed to dioxus ssr
