--- api/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "api"
-version = "0.2.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-tokio = { version = "1.38.0", features = ["full"] }
-shared = { path = "../shared" }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tracing = "0.1.40"
-mime = "0.3.17"
-axum = "0.7.5"
-tower-http = { version = "0.5.2", features = ["fs"] }
-tower = { version = "0.4.13", features = ["util"] }
