--- front/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "front"
-version = "1.0.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[profile.release]
-strip = true
-debug = false
-lto = true
-opt-level = "s"
-panic = "abort"
-
-[dependencies]
-dioxus = "0.4.3"
-dioxus-web = "0.4.3"
-log = "0.4.17"
-reqwest = { version = "0.12.5", features = ["json"] }
-shared = { path = "../shared" }
