--- front/public/styles/button.css
@@ -1,16 +1,0 @@
-.button {
-  padding: 8px 16px;
-  border: none;
-  border-radius: 4px;
-  background-color: #ffbf00;
-  color: #344B55;
-  font-size: 16px;
-  font-weight: bold;
-  cursor: pointer;
-  outline: none;
-  transition: ease all 0.2s;
-}
-
-.button:hover {
-  background-color: #ffdc73;
-}
