--- front/public/styles/field.css
@@ -1,23 +1,0 @@
-.field {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-.field.radio {
-  flex-direction: row-reverse;
-  cursor: pointer;
-  margin-bottom: 4px;
-}
-
-.field.free-text {
-  margin-right: 10px;
-  flex-direction: row;
-  align-items: center;
-}
-
-.field.free-text > input[type=text] {
-  margin: 0 20px 0 10px;
-  width: 100px;
-}
