--- front/public/styles/index.css
@@ -1,195 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: #CBC8C8;
-}
-
-main {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: stretch;
-  width: 100vw;
-  min-height: 100vh;
-  max-height: 100vh;
-  font-family: sans-serif;
-  overflow: hidden;
-}
-
-main > article.container {
-  flex: 1;
-  background-color: #CBC8C8;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: stretch;
-  padding: 20px 40px 40px;
-  box-sizing: border-box;
-  height: 100%;
-  max-width: 1200px;
-  width: 100%;
-  margin: auto;
-  overflow: hidden;
-}
-
-main > article.container > section {
-  flex: 1;
-  transition: ease-in-out all 0.2s;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: stretch;
-}
-
-main > article.container > section > header {
-  width: 100%;
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  min-height: 58px;
-}
-
-main > article.container > section svg {
-  max-width: 45px;
-  color: #344b55;
-}
-
-main > article.container > section > div {
-  flex: 1;
-  background-color: rgba( 52,75,85, 0.7);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  border: 1px solid rgba( 52,75,85, 0.5);
-  color: #CBC8C8;
-  border-radius: 6px;
-  box-shadow: rgba(0, 0, 0, 0.19) 0px 10px 20px, rgba(0, 0, 0, 0.23) 0px 6px 6px;
-  transition: ease all 0.2s;
-  padding: 20px 30px;
-  box-sizing: border-box;
-}
-
-main > article.container > section .scores {
-  flex: 0;
-}
-
-main > article.container > section .history {
-  margin-top: 20px;
-  flex: 1;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-}
-
-main > article.container > section .history > header {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: row;
-  margin-bottom: 20px;
-}
-
-main > article.container > section .history > header > .title {
-  margin: 0;
-}
-
-main > article.container > section .history ul {
-  margin: 0;
-  padding: 0;
-  overflow: auto;
-  overflow-x: hidden;
-}
-
-main > article.container > section.competitor {
-  margin-right: 40px;
-}
-
-/* main > article.container > section:hover > div {
-  box-shadow: rgba(0, 0, 0, 0.3) 0px 19px 38px, rgba(0, 0, 0, 0.22) 0px 15px 12px;
-  transform: scale(1.008);
-} */
-
-main > article.container .title {
-  margin: 0 0 20px;
-}
-
-main > article.container > section > div > .row {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-main > article.container > section > div > .row > .column {
-  flex: 0 1 auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  margin: 0 20px 20px 0;
-}
-
-main > article.container > section > div > .row > .column.full-width {
-  width: 100%;
-}
-
-main > article.container > section > div > .row > .column > .content {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-main > article.container > section > div input[type=text] {
-  padding: 6px 8px;
-  box-sizing: border-box;
-  border: 2px solid transparent;
-  border-radius: 2px;
-  outline: none;
-  transition: ease all 0.2s;
-  font-size: 14px;
-}
-
-main > article.container > section > div input[type=text].invalid {
-  border-color: red;
-  color: red;
-}
-
-@media only screen and (max-width: 950px) {
-  main {
-    overflow: auto;
-    max-height: none;
-  }
-
-  main > header {
-    padding: 10px 20px;
-  }
-
-  main > header > div > h1 {
-    margin: 0 20px;
-    font-size: 1.4em;
-  }
-
-  main > article.container {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    padding:  0 20px 40px;
-    box-sizing: border-box;
-  }
-
-  main > article.container > section.competitor {
-    margin-right: 0;
-  }
-
-  main > article.container > section svg {
-    max-width: 40px;
-    color: #355b55;
-  }
-}
