--- front/public/styles/score.css
@@ -1,68 +1,0 @@
-.score-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  flex-wrap: wrap;
-  padding: 20px;
-  box-sizing: border-box;
-  border: 1px solid #CBC8C833;
-  border-radius: 4px;
-  margin-bottom: 10px;
-}
-
-
-.score-container:last-of-type {
-  margin-bottom: 10px;
-}
-
-.score-row {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  flex-wrap: wrap;
-  overflow: hidden;
-}
-
-.score-competitor {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: flex-start;
-  margin-bottom: 20px;
-}
-
-.score-competitor h4 {
-  margin: 0;
-  flex: 1;
-}
-
-.score-competitor aside {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-end;
-  font-size: 14px;
-}
-
-.score-bubble {
-  padding: 10px;
-  box-sizing: border-box;
-  border-radius: 4px;
-  font-size: 16px;
-  background-color: #CBC8C8;
-  color: #344B55;
-  margin: 0 8px 8px 0;
-}
-
-.score-bubble:last-of-type {
-  margin-right: 0;
-}
-
-.score-bubble > span {
-  font-size: 18px;
-  font-weight: bold;
-  margin-right: 8px;
-}
