--- shared/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "shared"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = { workspace = true }
-serde_json = { workspace = true }
