--- shared/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod helpers;
-pub mod models;
