+++ src/components/score_bubble.rs
@@ -1,6 +1,7 @@
+#![allow(non_snake_case)]
 use dioxus::prelude::*;
 
-#[inline_props]
+#[component]
 pub fn ScoreBubble<'a>(cx: Scope<'a>, label: &'a str, score: &'a f64) -> Element {
     cx.render(rsx!(p {
         class: "score-bubble",
