+++ src/components/title.rs
@@ -1,6 +1,7 @@
+#![allow(non_snake_case)]
 use dioxus::prelude::*;
 
-#[inline_props]
+#[component]
 pub fn Title<'a>(cx: Scope<'a>, children: Element<'a>) -> Element {
     cx.render(rsx!(h2 {
         class: "title",
