+++ src/models/competitor.rs
