--- static/assets/dioxus/snippets/dioxus-web-a95d8cc6c91ff8eb/inline0.js
@@ -1,11 +1,0 @@
-
-    export function get_form_data(form) {
-        let values = new Map();
-        const formData = new FormData(form);
-
-        for (let name of formData.keys()) {
-            values.set(name, formData.getAll(name));
-        }
-
-        return values;
-    }
